Fix spot pricing bug
If the spot price for a particular instance type was close to half of
the on-demand instance price, then determine_best_price() would
incorrectly attempt to bid half of the on-demand price, even though that
could be lower than the current spot price. This could lead to clusters
not being provisioned with any instances.