Fix spot pricing logic
Fixed a bug where the spot pricing logic in get_bid_price() would base
the bid on the AZ which currently provides the lowest instance price.

However, if an EC2 Subnet ID is provided, then we can only bid on
instances that will be launched in the Availability Zone that is
associated with the provided EC2 Subnet ID, which means that prices in
other subnets should be ignored.